refactor(core)!: remove redundant payer-signing and Permit2 exports
