+++ packages/core/src/payment/index.ts
@@ -1,24 +1,6 @@
 export {
-  type SignReceiveAuthorizationParameters,
-  type SignReceiveAuthorizationReturnType,
-  signReceiveAuthorization,
-} from './erc3009.js'
-export {
-  type ComputeEscrowNonceReturnType,
   type ComputePaymentInfoHashReturnType,
-  computeEscrowNonce,
   computePaymentInfoHash,
   PAYMENT_INFO_TYPEHASH,
 } from './hashing.js'
-export {
-  type CreatePermit2ApprovalTxReturnType,
-  createPermit2ApprovalTx,
-  type GetPermit2AllowanceReadParamsInput,
-  type GetPermit2AllowanceReadParamsReturnType,
-  getPermit2AllowanceReadParams,
-  PERMIT2_ADDRESS,
-  type SignPermit2AuthorizationParameters,
-  type SignPermit2AuthorizationReturnType,
-  signPermit2Authorization,
-} from './permit2.js'
 export { validatePaymentInfo } from './validation.js'
