+++ packages/sdk/src/index.ts
@@ -7,7 +7,6 @@
 // ---------------------------------------------------------------------------
 export type {
   ConditionSlot,
-  CreatePermit2ApprovalTxReturnType,
   EvidenceEntry,
   FeeAddresses,
   FeeCalculationResult,
@@ -15,18 +14,12 @@ export type {
   GetEvidenceBatchReturnType,
   GetOperatorRefundRequestsReturnType,
   GetPayerRefundRequestsReturnType,
-  GetPermit2AllowanceReadParamsInput,
-  GetPermit2AllowanceReadParamsReturnType,
   GetReceiverRefundRequestsReturnType,
   OperatorConfig,
   OperatorSlots,
   PaymentAmounts,
   PaymentInfoWire,
   RefundRequestData,
-  SignPermit2AuthorizationParameters,
-  SignPermit2AuthorizationReturnType,
-  SignReceiveAuthorizationParameters,
-  SignReceiveAuthorizationReturnType,
   SupportedChainId,
   X402rChainConfig,
 } from '@x402r/core'
@@ -37,18 +30,13 @@ export {
   ConfigError,
   ContractCallError,
   computePaymentInfoHash,
-  createPermit2ApprovalTx,
   formatError,
   formatFeeBreakdown,
   getChainConfig,
-  getPermit2AllowanceReadParams,
   isSupportedChain,
   PaymentInfo,
-  PERMIT2_ADDRESS,
   RefundRequestStatus,
   SubmitterRole,
-  signPermit2Authorization,
-  signReceiveAuthorization,
   supportedChainIds,
   ValidationError,
   validateFeeBounds,
